Fix func convert_permalink_sign() to default
For documents themselves, this function is redundant.
because sample.t.css includes a cleaner solution.
(just 'display: none;')

But sadly it can't delete permalinks in pdf bookmarks.
Bookmarks use element text as node label
and document css' are not applied.

And as long as I can't think up a better way to delete them,
this function should be in defaults.