Add class Sites
And fixed new errors
in pdfname related code and some tests.
(Now .sites and .urls are generators, not lists,
we can't index them like sites[0] or urls[0].)

Also I had to temporarily change pdfname spec.
Now pdfname when urls is None raises Error,
and it is not desirable,
since it is better to be able to run '--appcheck' without urls.