Add 'smarter' very_short_run
It does slower converter tests conditionally,
when it seems suspicious, with very rough heuristics.
So I'm beginning to hesitate even to call it a 'test'.

Let's call it a psychological confidence amplifier.

Anyway, I believe how partial and terrible they may be,
faster and automatic tests bring a better result in the end,
when all things are considered.