Fix bug, sample url in tox env directory
We have to update the ufile copy in /tests/temp/actualrun,
if the timestamp of the original is newer,
even when running in tox environment.
But which would make a url like:
'[...]/.tox/py37/lib/python3.7/site-packages/tosixinch/templite.py',
since it resolve the path relative to the installed directory.

I cannot prevent this, and it is not in itself 'not functional',
but it is better I rewrite this to the ordinary path,
as soon as running in an ordinary test environment.