Cut relative templite.py in sample, use remote textwrap.py [API]
I wanted to include some relative filepath in test,
but It made the code unnecessarily complicated.

I was just testing a combination of a few os.path functions anyway.
And to be able to do so,
I needed to build even more combination of os.path functions.
It was not making much sense.

The selection of textwrap (among stdlib) is arbitrary
(short, seems to have long lines, etc.).