Fix unconditional prefix './' for relative url [API]
I want to remove '_in_current_dir' check in location.py.
It is better to be only used in system.py (os.path complication).

For this, 'component_url' was rewritten,
in the process, unconditional './' addition was removed.

It was not in itself wrong here,
but not so clean. (cf. '../aaa' -> './../aaa').
And the logic was greatly simplified.

But... In the process,
I had to add special case handling.
(Colon in first path in relative url is illegal.)

>>> urllib.parse.split('aa:bb/cc')
SplitResult(scheme='aa', netloc='', path='bb/cc',...