Fix unquoted url references in webpages
For user input urls,
The script doesn't do anything.

But some webpages have unquoted references like:

    <img src="pic 1.jpg">

They are not in user's control,
and the script aborts with `InvalidURL` error.

So only for `Component`,
the path part of input url is quoted.

Note reserved characters are excluded in quoting.
(they are handled in _tamper_* funtions).