Add process/sample.py (Cut util.py, gen.py, site.py)
Combine three files into one

This is a mechanical concatenation,
in the order of util, gen, site,
with the additional line of '# flake8: noqa'.

So the file is not yet a correct Python module.

---

I tried a history-keeping merge, with 'git commit-tree',
following Raymond Chen's blog article.
https://devblogs.microsoft.com/oldnewthing/20190515-00/?p=102495

But It introduced extra complexities for the repository editing
('git rebase' etc.),
I decided I can't afford other than the simplest git commits.